# gatewayservice/openapi.yaml
openapi: 3.0.0
info:
  title: Gateway Service API
  version: 1.0.0
  description: API gateway for managing authentication, user information, game statistics, and questions.
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://51.142.183.63:8000
    description: Production server
paths:
...
  /stats:
    get:
      summary: Get Statistics
      parameters:
        - in: query
          name: user
          schema:
            type: string
          description: Username for which to retrieve statistics
        - in: query
          name: gamemode
          schema:
            type: string
            enum: [clasico, bateria, calculadora]
          description: Game mode for which to retrieve statistics
      responses:
        "200":
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
              example:
                username: admin
                gamemode: clasico
                nGamesPlayed: 17
                avgPoints: 0.5882352941176471
                totalPoints: 10
                totalCoorectQuestions: 10
                totalIncorrectQuestions: 156
                ratioCorrect: 6.024096385542169
                avgTime: 10
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: Error al obtener las estadísticas
  /saveGame:
    post:
      summary: Save Game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Game"
      responses:
        "200":
          description: Game saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /ranking:
    get:
      summary: Get Ranking
      parameters:
        - in: query
          name: gamemode
          schema:
            type: string
            description: Game mode for which to retrieve ranking
        - in: query
          name: filterBy
          schema:
            type: string
            enum: [avgPoints, totalPoints, ratioCorrect, avgTime, avgPoints]        
      responses:
        "200":
          description: Ranking retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingList"
              example:
                - username: test1
                  avgPoints: 0.5882352941176471
                - username: test2
                  avgPoints: 0.6782352941176471
                - username: test3
                  avgPoints: 0.1232352941176471
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: Error al obtener el ranking
components:
  schemas:
...
    Game:
      type: object
      properties:
        gameMode:
          type: string
          enum: [clasico, bateria, calculadora]
        correctAnswers:
          type: integer
        incorrectAnswers:
          type: integer
        points:
          type: integer
        avgTime:
          type: number
        _id:
          type: string  
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
    Question:
      type: object
      properties:
        pregunta:
          type: string
          description: Question text
        respuestas:
          type: array
          description: List of possible answers
          items:
            type: string
        correcta:
          description: Correct answer
          type: string
        respuesta:
          description: User's answer
          type: string
    Stats:
      type: object
      properties:
        username:
          type: string
        gamemode:
          type: string
        nGamesPlayed:
          type: integer
        avgPoints:
          type: number
        totalPoints:
          type: integer
        totalCoorectQuestions:
          type: integer
        totalIncorrectQuestions:
          type: integer
        ratioCorrect:
          type: number
        avgTime:
          type: number
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
    RankingList:
      type: array
      items:
        type: object
        properties:
          username:
            type: string
          avgPoints:
            optional: true
            type: number
          totalPoints:
            optional: true
            type: integer
          ratioCorrect:
            optional: true
            type: number
          avgTime:
            optional: true
            type: number
    Error:
      type: object
      properties:
        error:
          type: string
...